chore: adapt ValidationHook signature TECH-880
So that a hook only calls the reporter in case it has to add an error.
This should allow us to get rid of the merging logic that is only needed
due to the signature of returning a ValidationErrorReporter

We might be able to remove the TrackerImportValidationContext and only
pass the TrackerBundle into Validators as is hinted at by the TODO
comment.