chore: hooks should add errors via reporter directly
my vision for the validation would be that a hook only adds errors, warning directly to
the ValidationErrorReporter. Ideally even only to the TrackerValidationReport and that
the ValidationErrorReporter doesnt even exist. No need for merging different ValidationErrorReporter
which was only needed because each Hook returned a ValidationErrorReporter instead of getting it
passed in. The code is still too intertwined, so I need to slowly untangle it and remove
dependencies.